django-backend/requirements.txt
asgiref==3.8.1
Django==5.0.7
django-prometheus==2.3.1
djangorestframework==3.15.2
mysqlclient==2.2.4
prometheus_client==0.20.0
sqlparse==0.5.0
